Forces CD-ROM door unlocked when umount is called
For systems that use systemd, it will not allow the ability to have a
cdrom door unlocked, which makes automating the unmounting of boot ISOs
after VM creation difficult. VMware will prompt the user on how to
handle the situation, during this time the VM is in a suspended,
unresponsive, state. This commit makes a string match guess on when the
question is asked, and forces a "Yes" response to VMware to unlock the
drive.

See https://github.com/systemd/systemd/pull/4643 for more info.