Use API for mkbootiso + plugins
This removes the plugins framework (which was only used by mkbootiso)
and uses a Flask API framework (1ef44994) for easier interaction with
this feature and others.

The former plugin, mkbootiso, now makes API calls to build ISOs on a per
server basis. The "templates" option was awkward and removed. The new
method is to merge an OS default dictionary in vctoolsrc.yaml with what
is provided in server.yaml. If the key under the defaults option matches
its "guestId", then it will merge the configs together with the
server.yaml taking precedence. This is only supported on systems that
boot using Anaconda.