Fix disk layout on VM creation
In the past, the only configuration was up to 4 disks each on their own
SCSI adapter. This commit allows for a new disk structure to add many
more disks. The new layout is a dictionary were the scsi number is the
key, and the value is a list of disk sizes:

  # sample.yaml
  ---
  vmconfig:
    disks:
      0:
      - 50
      - 10
      1:
      - 25

The old layout of 4 disks on their own SCSI adapter will still function
normally if disks is list.