Fix Travis checks with 2to3 conversion
The system package for mod_wsgi is compiled against 3.4 on the "trusty"
instance. For any later version it needs to be installed using pip and
the dev instances of python and apache2 to pass the checks. Also, this
removes the commit_check.py script since it is failing tests and not
compatible with the conversion.