Add options when key has no value
Allows for keys with no value to be used on the kickstart boot line.
Useful for options that act as flags, like "inst.noverifyssl". In the
yaml config, you can omit the value by using null. False also works, but
is misleading since it isn't disabling anything, but rather the
opposite.

    inst.noverify: null